Refactor Cypress tests
