+++ frontend/cypress/integration/layers_sidebar.spec.js
@@ -39,7 +39,7 @@ context('LayersSidebar', () => {
     // Show a zone with the zone button
     cy.log('Show a zone with the zone button')
     cy.get('*[data-cy="regulatory-layers-my-zones-zone-show"]').eq(0).click({ timeout: 20000 })
-    cy.wait(1000)
+    cy.wait(500)
 
     cy.get('canvas', { timeout: 20000 }).eq(0).click(490, 580, { timeout: 20000, force: true })
     cy.get('*[data-cy="regulatory-layers-metadata-lawtype"]').contains('Reg. MEMN')
@@ -73,7 +73,7 @@ context('LayersSidebar', () => {
     // Show a zone with the topic button
     cy.log('Show a zone with the topic button')
     cy.get('*[data-cy="regulatory-layers-my-zones-topic-show"]').eq(0).click({ timeout: 20000 })
-    cy.wait(1000)
+    cy.wait(500)
 
     cy.get('canvas').eq(0).click(490, 580, { timeout: 20000, force: true })
     cy.get('*[data-cy="regulatory-layers-metadata-lawtype"]').contains('Reg. MEMN')
@@ -184,7 +184,7 @@ context('LayersSidebar', () => {
     cy.get('*[data-cy^="layers-sidebar"]').click({ timeout: 20000 })
     cy.get('*[data-cy^="administrative-zones-open"]').click({ timeout: 20000, force: true })
     cy.get('*[data-cy^="administrative-layer-toggle"]').eq(0).click({ timeout: 20000 })
-    cy.wait(1000)
+    cy.wait(500)
 
     // Then
     cy.get('.administrative').eq(0).toMatchImageSnapshot({
