+++ frontend/cypress/integration/tritonfish.spec.js
@@ -8,7 +8,7 @@ context('TritonFish', () => {
     cy.visit(`http://localhost:${port}/ext#@-824534.42,6082993.21,8.70`)
     cy.get('*[data-cy^="first-loader"]', { timeout: 20000 }).should('not.exist')
     cy.url().should('include', '@-82')
-    cy.wait(1000)
+    cy.wait(200)
   })
 
   it('view Should have some features removed', () => {
