+++ frontend/cypress/integration/vessel_sidebar_identity.spec.js
@@ -1,0 +1,28 @@
+/* eslint-disable no-undef */
+/// <reference types="cypress" />
+
+const port = Cypress.env('PORT') ? Cypress.env('PORT') : 3000
+
+context('Vessel sidebar identity tab', () => {
+  beforeEach(() => {
+    cy.viewport(1280, 1024)
+    cy.visit(`http://localhost:${port}/#@-824534.42,6082993.21,8.70`)
+    cy.get('*[data-cy^="first-loader"]', { timeout: 20000 }).should('not.exist')
+    cy.url().should('include', '@-82')
+    cy.wait(200)
+  })
+
+  it('Identity Should contain the vessel identity', () => {
+    // Given
+    cy.get('.vessels').click(460, 480, { timeout: 20000, force: true })
+    cy.get('*[data-cy^="vessel-sidebar"]', { timeout: 20000 }).should('be.visible')
+
+    // When
+    cy.get('*[data-cy^="vessel-menu-identity"]', { timeout: 20000 }).should('be.visible')
+    cy.get('*[data-cy^="vessel-menu-identity"]').click({ timeout: 20000 })
+
+    // Then
+    cy.get('*[data-cy^="vessel-identity-gears"]', { timeout: 20000 }).should('be.visible')
+    cy.get('*[data-cy^="vessel-identity-gears"]').contains('Sennes danoises (SDN)', { timeout: 20000 })
+  })
+})
