+++ frontend/cypress/integration/vessel_visibility.spec.js
@@ -19,7 +19,7 @@ context('Vessel visibility', () => {
     cy.get('*[data-cy="map-property-trigger"]')
       .filter(':contains("les navires au port")')
       .click({ timeout: 20000, force: true })
-    cy.wait(1000)
+    cy.wait(500)
     cy.get('#root').dblclick(560, 620, { timeout: 20000 })
 
     // Then
